Create perfect_number.c++
A positive integer p is called a perfect number if all the proper divisors of p sum to p exactly. Integer d is a proper divisor of p if 1 <= d <= p - 1 and p is evenly divisible by d. For example, the number 28 is a perfect number, since its proper divisors (which are 1, 2, 4, 7, 14) add up to 28. Write a program that repeatedly asks the user to enter an integer. And, then the program tests the given integer is perfect number or not and displays the result. Use a do loop that asks the user whether the program should be terminated. Typical interaction might be
Enter an integer: 6
6 is perfect number.
Do you wish to continue (y/n)? y
Enter an integer: 7
7 is not perfect number.
Do you wish to continue (y/n)? n