Create student.c++
Question:

Create a structure called student that contains a name (char array), and roll number (type long).  Assume the name has no embedded blanks. Write a main() program to exercise this structure. It should create an array of type student, and then invite the user to input data for up to 100 students.  Finally, it should print out the data for all the students.