Create Sort Integers.c++
Question:

Write a program that input 10 integers from user and store them in array. Then sort the array in ascending order and display the sorted result. When it finishes, the program should repeatedly ask whether the user wants to do another data to sort. The response can be ‘Y’ or ‘N’. If user enter ‘N’, at which point the program terminates. Use do-while loop.